LayoutTests/media/media-source/media-source-real-webm-discontinuity-expected.txt


EVENT(sourceopen)
EVENT(updateend)
EVENT(updateend)
EXPECTED (sourceBuffer.buffered.length == '1') OK
EVENT(updateend)
EXPECTED (sourceBuffer.buffered.length == '2') OK
EXPECTED (sourceBuffer.buffered.start(0) == '0') OK
EXPECTED (sourceBuffer.buffered.end(0) == '2') OK
EXPECTED (sourceBuffer.buffered.start(1) == '10') OK
EXPECTED (sourceBuffer.buffered.end(1) == '12') OK
END OF TEST
